update scroll buttons
